Removed unused utilities and made one data frame to use for all tests
