+++ tests/test_utils.py
@@ -7,51 +7,6 @@ import iup
 import iup.utils
 
 
-@pytest.fixture
-def frame():
-    """
-    Make a mock data frame to test model building.
-    """
-    frame = pl.DataFrame(
-        {
-            "geography": ["USA", "PA", "USA", "PA", "USA", "PA", "USA", "PA"],
-            "time_end": [
-                "2019-12-31",
-                "2019-12-31",
-                "2020-01-07",
-                "2020-01-07",
-                "2020-01-14",
-                "2020-01-14",
-                "2020-01-21",
-                "2020-01-21",
-            ],
-            "estimate": [0.0, 0.0, 1.0, 0.1, 3.0, 0.3, 4.0, 0.4],
-            "season": "2019/2020",
-            "elapsed": [0, 0, 7, 7, 14, 14, 21, 21],
-            "interval": [None, None, 7, 7, 7, 7, 7, 7],
-        }
-    )
-
-    frame = frame.with_columns(time_end=pl.col("time_end").str.to_date("%Y-%m-%d"))
-
-    return frame
-
-
-def test_extract_standards(frame):
-    """
-    Make a dictionary with as many keys as provided variables,
-    and two values per sub-dictionary
-    """
-    output = iup.utils.extract_standards(frame, ("estimate", "elapsed"))
-
-    correct = {
-        "estimate": {"mean": frame["estimate"].mean(), "std": frame["estimate"].std()},
-        "elapsed": {"mean": frame["elapsed"].mean(), "std": frame["elapsed"].std()},
-    }
-
-    assert output == correct
-
-
 def test_date_to_season(frame):
     """
     Season is extracted from a column of dates
@@ -98,19 +53,6 @@ def test_date_to_elapsed_handles_season_start(frame):
     )
 
     assert (output["elapsed"] == expected).all()
-
-
-def test_date_to_interval(frame):
-    """
-    Return the interval between dates by grouping factor.
-    """
-    output = frame.sort(["geography", "time_end"]).with_columns(
-        interval=iup.utils.date_to_interval(pl.col("time_end")).over("geography")
-    )
-
-    expected = pl.Series("interval", ([None] + [7.0] * 3) * 2)
-
-    assert (output["interval"] == expected).all()
 
 
 @pytest.fixture
